Add files via upload
The Tower of Hanoi is a mathematical puzzle that demonstrates the power of recursion. The goal is to move a stack of disks of different sizes from a source rod to a destination rod, using an auxiliary rod, with the following rules:
Only one disk can be moved at a time.
Each move consists of taking the upper disk from one stack and placing it on top of another stack. 
No disk may be placed on top of a smaller disk.